collectors/python-ipmi-collector/requirements.txt
pyghmi==1.5.16
psycopg2-binary==2.9.9
requests==2.31.0
prometheus-client==0.19.0
python-dotenv==1.0.0
schedule==1.2.1
structlog==23.2.0
cryptography==41.0.7
